requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
numpy==1.18.1
pkg-resources==0.0.0
PySocks==1.7.1
python-dotenv==0.10.5
requests==2.22.0
stem==1.8.0
torrequest==0.1.0
urllib3==1.25.8
